fix(test): restore e2e curator-pipeline test clobbered by #1159
#1159 added a new curator-pipeline *scheduling* test but named the file
test_curator_pipeline.c — silently overwriting the pre-existing end-to-end curator
chain test (resolve_entities -> index_code_unit -> link_artifacts -> /v1/implements)
and leaving a duplicate Rules.mk recipe for unit-test-curator-pipeline (a build
warning; the two rules raced to define the same target).

Restore the original e2e test to test_curator_pipeline.c, move the scheduling test
to test_curator_pipeline_sched.c (unit-test-curator-pipeline-sched), and drop the
duplicate target/rule. Both tests build and pass.
